+++ app/src/main/java/com/mnnit/moticlubs/domain/model/Admin.kt
@@ -6,7 +6,7 @@ import androidx.room.Entity
 @Entity(tableName = "admin", primaryKeys = ["uid", "cid"])
 data class Admin(
     @ColumnInfo(name = "uid", index = true)
-    val userID: Int,
+    val userID: Long,
 
     @ColumnInfo(name = "cid", index = true)
     val clubID: Int
